Hex Math, and 2 New Effects
Big thanks again to naroa1 for the rebase of my spaghetti code. I added in some new functions to calculate addresess so that I won't need to enter in 300 addresses next time Konami updates the game. The code now just needs the addresses for Life Medicine and the Survival knife to have all the weapons,items and camos/face paints working. Still working at getting the same thing done for the MiscForm's effects. Looking into a way to future proof further with Signature scans (This is a new concept to me so still a work in progress and commented out atm in the code). I also added in some effects for infinite alert and caution for testers to test out before this version goes on Nexus Mods.