Area Strings and The Fear Slider
Constants.cs has a list of strings now that are used for different areas these will be used for future conditional checks. The Fear now has a slider to adjust his health and stamina. Will setup an area string check for him so memory can't be altered outside of his boss battle.