Clickonce Complete and Boss Refactor Underway
Clickonce is fully implemented the Setup.exe should be all an end user needs to boogy with the trainer now. Would like to get the files working with Github in the future but the current cloud host will work for now.

My new way of getting to the dynamic address without scanning 30 million potential addresses seems to be working for Ocelot and The Pain fingers crossed this will be a good fix for multiple versions.