Big Refactor
Changed up how I scan for AOBs and cut out like 1000ish lines of code in the process. The AOBs in Constants.cs are now in AobManager.cs with an extra field added to specify the range of the AOB when declaring the AOB. Makes things a lot cleaner and easier to manage/add new things in. I can probably add in the invincible guards next with no issues now thanks to this. I also added in some textboxes to allow users to change Snake's position.