Camo Index Fixes
A user had opened a ticket on Discord about Camo Index changes not working properly due to the camo changing back to default when the game was paused or a new area was entered. Also a big oversight on my part was mention I forgot to list somewhere where the version was visible so users know what version they are on so each form title states the newest version of the trainer now. Also added in a TimerManager class for managing checkboxes easier in the future.