Fixed Broken Effects + New Effects
MGS3 version 2.0.0 was released a couple weeks a go the fixes were already pushed to the ClickOnce when that happened. This update to the repo has that and a few new features I've been looking to implement.

New Features:
Time of Day Changes
Randomizing Guard Locations
Better way to locate the main pointer with Snake's health
Much Faster way of locating Snake and Guard positions